chore(hooks): second round code review changes
